web/requirements.txt
Flask
Werkzeug
Flask_Mail
Flask-Login
Flask-MySQLdb
mysql-connector-python
Flask-SQLAlchemy
Flask-Migrate
nose2 # test runner
Flask-Fixtures
PyYAML
Flask-Script
